requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
filelock==3.8.0
huggingface-hub==0.10.0
idna==3.4
importlib-metadata==5.0.0
numpy==1.21.6
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.1
transformers==4.22.2
typing-extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1